move everything from /page-layout to /components-layout
