+++ products/statement-generator/src/components-layout/Affirmation.tsx
