+++ products/statement-generator/src/components-layout/AppFooter.tsx
