+++ products/statement-generator/src/pages-form/InvolvementCommunityServiceFlow.tsx
@@ -4,9 +4,9 @@ import FormStateContext from 'contexts/FormStateContext';
 
 import Textarea from 'components/Textarea';
 
-import ContentContainer from 'page-layout/ContentContainer';
-import FlowNavigation from 'page-layout/FlowNavigation';
-import FormContainer from 'page-layout/FormContainer';
+import ContentContainer from 'components-layout/ContentContainer';
+import FlowNavigation from 'components-layout/FlowNavigation';
+import FormContainer from 'components-layout/FormContainer';
 
 import Input from '../components/Input';
