+++ products/statement-generator/src/pages-form/WhyStep.tsx
@@ -5,9 +5,9 @@ import FormStateContext from 'contexts/FormStateContext';
 import HelpPopUp from 'components/HelpPopUp';
 import Textarea from 'components/Textarea';
 
-import ContentContainer from 'page-layout/ContentContainer';
-import FlowNavigation from 'page-layout/FlowNavigation';
-import FormContainer from 'page-layout/FormContainer';
+import ContentContainer from 'components-layout/ContentContainer';
+import FlowNavigation from 'components-layout/FlowNavigation';
+import FormContainer from 'components-layout/FormContainer';
 
 function WhyStep() {
   const { formState, updateStepToForm } = useContext(FormStateContext);
