+++ products/statement-generator/src/pages/AboutUs.tsx
@@ -4,7 +4,7 @@ import { makeStyles, createStyles, Link } from '@material-ui/core';
 
 import Button from 'components/Button';
 
-import ContentContainer from 'page-layout/ContentContainer';
+import ContentContainer from 'components-layout/ContentContainer';
 
 import useUtilityStyles from 'styles/utilityStyles';
