+++ products/statement-generator/src/pages/PrivacyPolicy.tsx
@@ -3,7 +3,7 @@ import { useHistory } from 'react-router-dom';
 
 import Button from 'components/Button';
 
-import ContentContainer from 'page-layout/ContentContainer';
+import ContentContainer from 'components-layout/ContentContainer';
 
 import useUtilityStyles from 'styles/utilityStyles';
