+++ products/statement-generator/src/pages/Welcome.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import ContentContainer from 'page-layout/ContentContainer';
-import FlowNavigation from 'page-layout/FlowNavigation';
+import ContentContainer from 'components-layout/ContentContainer';
+import FlowNavigation from 'components-layout/FlowNavigation';
 
 export default function Welcome() {
   return (
